// lib/api/generated/models/aiPlanRecipe.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * CKS Slavia API
 * API panelu klubowego CKS Slavia Ruda Śląska
 * OpenAPI spec version: 2.2.0.1+1
 */
import type { AiCycleDay } from './aiCycleDay';
import type { AiPhase } from './aiPhase';
import type { AiWave } from './aiWave';

export interface AiPlanRecipe {
  cycle?: AiCycleDay[];
  /**
     * @minimum 0
     * @nullable
     */
  meet_week?: number | null;
  /** @nullable */
  notes?: string | null;
  phases?: AiPhase[];
  /** @nullable */
  title?: string | null;
  waves?: AiWave[];
  /**
     * @minimum 0
     * @nullable
     */
  weeks?: number | null;
}